Move test into separate useObserver test file
I've tweaked the test name to reflect what it's really checking for.
